+++ src/pages/PromotionPage/PromotionPage.tsx
@@ -1,54 +1,18 @@
-import { useRef, useState } from 'react';
-
 import './PromotionPage.css';
 
 import { promotions } from '../../data';
 
-import { PROMOTION } from '../../constants';
+import { usePromotion } from '../../hooks/usePromotion';
 
 import PromotionCard from '../../components/PromotionCard/PromotionCard';
 
 function PromotionPage() {
-  const promotionListRef = useRef<HTMLUListElement>(null);
-
-  const [buttonDisabled, setButtonDisabled] = useState({
-    prev: true,
-    next: false,
-  });
-
-  const handleScrollPromotionList = () => {
-    const promotionList = promotionListRef.current;
-
-    if (!(promotionList instanceof HTMLUListElement)) return;
-
-    setButtonDisabled({
-      prev: !promotionList.scrollLeft,
-      next:
-        promotionList.scrollLeft ===
-        promotionList.scrollWidth - promotionList.offsetWidth,
-    });
-  };
-
-  const handleClickControlButton = (direction: string) => {
-    const promotionList = promotionListRef.current;
-
-    if (!(promotionList instanceof HTMLUListElement)) return;
-
-    const offset =
-      direction === 'prev' ? -PROMOTION.MOVE_OFFSET : PROMOTION.MOVE_OFFSET;
-
-    promotionList.scrollBy({
-      left: offset,
-      behavior: 'smooth',
-    });
-
-    setButtonDisabled({
-      prev: !(promotionList.scrollLeft + offset),
-      next:
-        promotionList.scrollLeft + offset ===
-        promotionList.scrollWidth - promotionList.offsetWidth,
-    });
-  };
+  const {
+    buttonDisabled,
+    promotionListRef,
+    handleClickControlButton,
+    handleScrollPromotionList,
+  } = usePromotion();
 
   return (
     <div className='promotion-page'>
